Se imprime mensaje de error con comillas abiertas (status no manejado?), tester de lexing funcional

+++ parsing/heredoc2.c
@@ -41,13 +41,13 @@ int	heredoc_len(const char *line)
 	int		len;
 
 	i = 0;
-	while (i < ft_strlen(line) && i < 2 && ft_strchr(REDIR_S, line[i]))
+	while (i < (int)ft_strlen(line) && i < 2 && ft_strchr(REDIR_S, line[i]))
 		i++;
 	while (line[i] && ft_isspace(line[i]))
 		i++;
-	while (i < ft_strlen(line) && line[i])
+	while (i < (int)ft_strlen(line) && line[i])
 	{
-		if (i < ft_strlen(line) && ft_strchr(QUOTES, line[i])
+		if (i < (int)ft_strlen(line) && ft_strchr(QUOTES, line[i])
 			&& (i == 0 || (i > 0 && line[i - 1] != '\\')))
 		{
 			len = quoted_len(line + i, line[i]);
